clear up a bunch of windows warnings

+++ Code/DataStructs/ExplicitBitVect.h
@@ -41,7 +41,8 @@ class ExplicitBitVect : public BitVect {
   //! construct directly from a dynamic_bitset pointer
   //  takes ownership of the pointer
   ExplicitBitVect(boost::dynamic_bitset<> *bits)
-      : dp_bits(bits), d_size(bits->size()), d_numOnBits(bits->count()){};
+      : dp_bits(bits), d_size(static_cast<unsigned int>(bits->size())),
+        d_numOnBits(static_cast<unsigned int>(bits->count())){};
 
   ~ExplicitBitVect();
